+++ controller/message.go
@@ -27,7 +27,7 @@ func MessageAction(c *gin.Context) {
 
 	if user, exist := usersLoginInfo[token]; exist {
 		userIdB, _ := strconv.Atoi(toUserId)
-		chatKey := genChatKey(user.Id, int64(userIdB))
+		chatKey := genChatKey(int64(user.Id), int64(userIdB))
 
 		atomic.AddInt64(&messageIdSequence, 1)
 
@@ -55,7 +55,7 @@ func MessageChat(c *gin.Context) {
 
 	if user, exist := usersLoginInfo[token]; exist {
 		userIdB, _ := strconv.Atoi(toUserId)
-		chatKey := genChatKey(user.Id, int64(userIdB))
+		chatKey := genChatKey(int64(user.Id), int64(userIdB))
 
 		c.JSON(http.StatusOK, ChatResponse{Response: common.Response{StatusCode: 0}, MessageList: tempChat[chatKey]})
 	} else {
